applet.control.gpio: allow interacting with pins from CLI.
This is very useful to set strap pins, e.g. if you have a system reset
connected to pin `B7` and it must be asserted for JTAG to work:

    glasgow multi jtag-probe <...> ++ gpio --pins B7 B7=0

It is also possible to set pins to drive weakly (e.g. use the `B7=H` or
`B7=L` as an argument), and to sample pins (e.g. use `B6` as an
argument).