applet.interface.i2c_controller: use clock divisor API.
The way this is done is rather hacky, done so to avoid modifying
the existing I2C initiator as much as possible (since it is reliable
and also verifying I2C gateware is very difficult).

At some later point the existing I2C gateware will be replaced with
a stream-based one anyway. Then all of this will go away.