gateware.fx2_crossbar: flush IN FIFO once per microframe.
This eliminates a common pitfall for applet authors, where by default
the IN FIFO is flushed too often, and very small packets clog the FX2
buffers, causing buffer overflows and data loss. With this change, so
long as the host polls the device once per microframe (with XHCI hosts,
this can happen even more than once), there will never be a case where
an FX2 buffer isn't available, unless the bus is very heavily loaded
with other traffic.

There are two cases where more control over flushing is desired.

First, in request/response style protocols, it is useful to flush
the IN FIFO once the OUT FIFO is empty, since this reduces time spent
waiting for responses. This can significantly improve `probe-rs`
throughput, and is still used in that applet.

Second, an applet that produces a stream of data quickly enough that it
would be sent as a sequence of maximum-length packets may still need
this data to arrive at the host promptly. This can be an issue with
`uart` applet at high baud rates and duty cycles. It is not fixed in
this commit.

The throughput and latency when configured to flush implicitly, as
measured by the `benchmark` applet, are essentially unchanged;
the latency should be worse by ~125 µs (worst case), but this change
is comparable to noise.

All of the V2 applets that used to use explicit flush have been updated
to use autoflush where applicable; only `probe-rs` continues to use
explicit flush.

Fixes #263.
Fixes #794.