[WIP] gateware.fx2_crossbar: flush IN FIFO once per microframe.
This eliminates a common pitfall for applet authors, where by default
the IN FIFO is flushed too often, and very small packets clog the FX2
buffers, causing buffer overflows and data loss. With this change, so
long as the host polls the device once per microframe (with XHCI hosts,
this can happen even more than once), there will never be a case where
an FX2 buffer isn't available, unless the bus is very heavily loaded
with other traffic.

There is still one case where more control over flushing is desired:
an applet that produces a stream of data quickly enough that it would
be sent as a sequence of maximum-length packets, but needs this data to
arrive at the host promptly. To break up the data stream, `in_flush`
now has the behavior of asserting PKTEND on the 511th byte of a packet,
which causes the URB to immediately complete instead of continue to
fill up. This is rarely needed (only in the `uart` applet currently).

The throughput and latency, as measured by the `benchmark` applet, are
essentially unchanged; the latency may have gotten worse by ~125 µs,
but this is inconclusive.

All of the applets that used to use explicit flush have been updated
to use autoflush. The meaning of `in_flush=` has been changed from
"assert PKTEND the next time nothing is being written to the FIFO" to
"assert PKTEND the next time the FX2 FIFO contains exactly 511 bytes".
Where applicable applets have also been extended to use the new meaning.

Fixes #263.
Fixes #794.