applet.sensor.sen5x: add delay before I2C reads
The datasheet lists a worst-case command execution time of 20 ms for all
read commands, but doesn't clearly explain what to do with that information.

However, the device NACKs an I2C read if it's issued too soon after the
command-address write. So, this delays all reads by that execution time
and seems to be stable now.