applet.interface: Create the `uart-pinout` applet
Adds a uart-pinout applet that functions similar to the JTAGulator
implementation for determining UART pinouts. Essentially this just sends
data to the selected pins at various baud rates and looks for an echo.
The default just sends a carriage return (0x0d) if no data is set, but
the user can send arbitrary binary data up to 0xFF bytes long.