pyproject: constrain python-version in a less ambiguous way
`uv` pointed this out to me while running glasgow; the `~=` operator
for dependency resolution is a bit subtle; it basically allows you to
"round up" the upper bound of a dependency but it does so based on the
least-sigificant-specified component of a version number.

Basically that means `~=3.11` is equal to "from any version 3.11 up-to
< 4.0" but `~3.11.0` means "from any version 3.11.0 up-to < 3.12.0".

We want the first, but it's apparently subtle enough that uv (probably
correctly, tbh) warns about it. Let's write it in a more unambiguous
way.

Signed-off-by: Austin Seipp <aseipp@pobox.com>