firmware: don't bake GIT_REVISION into firmware
This certainly causes the firmware to be a different one on every
commit, and it'd become impossible to update the generated firmware
alongside (as the commit ID is CA and would change).

There's also no commit ID to determine if we're just dealing with a raw
archive of a specific revision, without a `.git` directory.

If we want to reintroduce some version ID into the firmware, it probably
should be something that doesn't change on each commit, and is
overridable externally, so build systems/packagers can pass the version
string externally.

Asking `pdm` for the version string [^1] might check these boxes; we can
define a less noisy version_format, it can be overridden with
PDM_BUILD_SCM_VERSION, and a fallback can be defined, but maybe not
baking one in at all for the sake of reproducibility is better?

[^1]: https://backend.pdm-project.org/metadata/