applet.interface.uart: change quit sequence to `~` then `.`.
This sequence is similar to SSH, except it doesn't have to follow
a newline. It's changed because on some non-US keyboard layouts it
may be difficult to enter Ctrl+\ (issue #998), and because it was
training users to use Ctrl+\ to quit other applets or modes, which
sends SIGQUIT and dumps core on some OSes (issue #568).

Fixes #998.