revC3: schematics, cosmetic: regenerate schematics.pdf
The new export is with KiCad 8 vs. the previously used 7. The project
has been updated to KiCad 8 a while back (in d884af4b), but the
schematic has never been re-exported since.

This means that there's a few changes that the schematic is now catching
up to:

  1. 3.3V is now called 3V3
  2. The symbols for resistor networks are slightly visually different.

Otherwise, however, the schematic is the same, and represents the same
netlist.