support.arepl: fix support for multi-line statements
The existing REPL supports multi-line statements by assuming that any line
it cannot compile must require more lines to make it complete; as soon as
it does get a valid parse then the code is evaluated or executed as
appropriate, even if the user wanted to provide further lines.

Instead, we can wait for the user to enter a blank line and treat that as
a positive indication that the multi-line statement is complete, allowing
for more complex definitions to be built up.