hardware.assembly: HardwareOutPipe fix buffer_size not working when > ~2000000
Previously, when buffer_size was set, send only checked the size of
inflight requests. This meant that if buffer_size was more than
~_max_packets_per_ep * 512 (2097152), send would never stall as the
size of inflight requests cannot exceed this value.

Changed to instead consider both the size of inflight requests and
the size of the contents of the buffer.