simulation.assembly: implement pulls and pin inversion for jumpers
While attempting to test a 24x I²C EEPROM emulator applet, I wanted to
hook it up to the `Memory24xInterface`. Since the I²C bus requires a
pull-up somewhere on the open-drain SCL/SDA lines, the
`Memory24xInterface` configures pull-ups for the pins it is connected
to. The simulation assembly however did not support this so far.
This led to a non-functioning bus.

`software/tests/gateware/test_i2c.py` manually sets the values of the
simulation ports' output signals based on the `oe` signals. For a more
general implemention for Glasgow simulation assemblies, this change
follows a similar pattern, setting jumper signals to the desired
`PullState`'s equivalent bit value, if all connected pins have their
output disabled.

While implementing this I noticed that pin inversion was being ignored
so far as well. This change also implements that, by passing the
requested inversion on to the created `SimulationPort`s.

With this change conflicting configuration requests for the same pin
lead to an exception. As do conflicting pulls on different pins
connected by a jumper.